refactor public submission program eligibility logic to use new service method and simplify tests
