+++ website/src/lib/services/program/program-public-submission.service.test.ts
@@ -1,10 +1,17 @@
 import { type PrismaClient } from '@/generated/prisma/client';
 import type { ServiceResult } from '../core/base.types';
-import { ProgramPublicSubmissionService } from './program-public-submission.service';
+import type { StoryblokService } from '../storyblok/storyblok.service';
+import { ProgramPublicSubmissionService, type PublicSubmissionProgramOption } from './program-public-submission.service';
 
 jest.mock('@/generated/prisma/client', () => ({
 	Prisma: {},
 	PrismaClient: class {},
+}));
+
+const mockFormatStoryblokUrl = jest.fn((filename: string) => `https://img.test/${filename}`);
+
+jest.mock('../storyblok/storyblok.utils', () => ({
+	formatStoryblokUrl: (...args: unknown[]) => mockFormatStoryblokUrl(...(args as [string])),
 }));
 
 const expectSuccess = <T>(result: ServiceResult<T>) => {
@@ -24,22 +31,173 @@ const expectFailure = (result: ServiceResult<unknown>, error: string) => {
 	expect(result.error).toBe(error);
 };
 
+type StoryblokProgram = {
+	content: {
+		portalSlug: string;
+		title: string;
+		description?: string;
+		primaryImage?: { filename: string; focus?: string };
+	};
+};
+
+const enProgram = (portalSlug: string, title: string): StoryblokProgram => ({
+	content: {
+		portalSlug,
+		title,
+		description: `${title} EN description`,
+		primaryImage: { filename: `${portalSlug}.jpg`, focus: '0x0:1x1' },
+	},
+});
+
+const deProgram = (portalSlug: string, title: string): StoryblokProgram => ({
+	content: {
+		portalSlug,
+		title,
+		description: `${title} DE description`,
+		primaryImage: { filename: `${portalSlug}-de.jpg`, focus: '0x0:1x1' },
+	},
+});
+
+const eligibleRow = {
+	id: 'program-1',
+	name: 'DB Name',
+	slug: 'si-core-sl',
+	countryId: 'country-sl',
+	countryIsoCode: 'SL' as const,
+	recipientsCount: 8,
+	focuses: [{ slug: 'poverty', name: 'Poverty' }],
+};
+
 const createService = () => {
 	const findMany = jest.fn();
 	const findFirst = jest.fn();
+	const getPrograms = jest.fn();
+	const getFocuses = jest.fn();
 	const db = {
 		program: {
 			findMany,
 			findFirst,
 		},
 	};
+	const storyblok = {
+		getPrograms,
+		getFocuses,
+	} as unknown as StoryblokService;
 	const logger = { error: jest.fn() };
-	const service = new ProgramPublicSubmissionService(db as unknown as PrismaClient, logger as never);
+	const service = new ProgramPublicSubmissionService(db as unknown as PrismaClient, storyblok, logger as never);
 
-	return { service, findMany, findFirst, logger };
+	return { service, findMany, findFirst, getPrograms, getFocuses, logger };
 };
 
 describe('ProgramPublicSubmissionService', () => {
+	describe('getEligibleProgramsForPublicSubmission', () => {
+		beforeEach(() => {
+			jest.clearAllMocks();
+		});
+
+		test('uses default-language Storyblok slugs for eligibility even when enriching another locale', async () => {
+			const { service, findMany, getPrograms, getFocuses } = createService();
+			getPrograms.mockImplementation(async (lang: string) => {
+				if (lang === 'en') {
+					return { success: true, data: [enProgram('si-core-sl', 'Core EN'), enProgram('only-en', 'Only EN')] };
+				}
+
+				return { success: true, data: [deProgram('si-core-sl', 'Core DE'), deProgram('only-de', 'Only DE')] };
+			});
+			getFocuses.mockResolvedValue({
+				success: true,
+				data: [{ content: { portalSlug: 'poverty', title: 'Poverty alleviation' } }],
+			});
+			findMany.mockResolvedValue([
+				{
+					id: 'program-1',
+					name: 'DB Name',
+					slug: 'si-core-sl',
+					countryId: 'country-sl',
+					country: { isoCode: 'SL' },
+					targetFocuses: [{ focus: { name: 'Poverty', slug: 'poverty' } }],
+					_count: { recipients: 8 },
+				},
+			]);
+
+			const data = expectSuccess(await service.getEligibleProgramsForPublicSubmission('de'));
+
+			expect(getPrograms).toHaveBeenCalledWith('en');
+			expect(getPrograms).toHaveBeenCalledWith('de');
+			expect(findMany).toHaveBeenCalledWith(
+				expect.objectContaining({
+					where: {
+						slug: { in: ['si-core-sl', 'only-en'] },
+						recipients: { some: {} },
+					},
+				}),
+			);
+			expect(data[0]).toMatchObject({
+				id: 'program-1',
+				name: 'Core DE',
+				description: 'Core DE DE description',
+				imageUrl: 'https://img.test/si-core-sl-de.jpg',
+				tags: ['Poverty alleviation'],
+			} satisfies Partial<PublicSubmissionProgramOption>);
+		});
+
+		test('falls back to DB name and null media when Storyblok enrichment is missing', async () => {
+			const { service, findMany, getPrograms, getFocuses } = createService();
+			getPrograms.mockResolvedValue({
+				success: true,
+				data: [enProgram('other-program', 'Other')],
+			});
+			getFocuses.mockResolvedValue({
+				success: true,
+				data: [{ content: { portalSlug: 'poverty', title: 'Poverty alleviation' } }],
+			});
+			findMany.mockResolvedValue([
+				{
+					...eligibleRow,
+					slug: 'missing-in-storyblok',
+					country: { isoCode: 'SL' },
+					targetFocuses: [{ focus: { name: 'Poverty', slug: 'poverty' } }],
+					_count: { recipients: 8 },
+				},
+			]);
+
+			const data = expectSuccess(await service.getEligibleProgramsForPublicSubmission('en'));
+
+			expect(data[0]).toMatchObject({
+				name: 'DB Name',
+				description: null,
+				imageUrl: null,
+				tags: ['Poverty alleviation'],
+			});
+		});
+
+		test('propagates eligibility service failures', async () => {
+			const { service, findMany, getPrograms, getFocuses, logger } = createService();
+			getPrograms.mockResolvedValue({ success: true, data: [enProgram('si-core-sl', 'Core EN')] });
+			getFocuses.mockResolvedValue({ success: true, data: [] });
+			findMany.mockRejectedValue(new Error('db down'));
+
+			const result = await service.getEligibleProgramsForPublicSubmission('en');
+
+			expectFailure(result, 'Could not load programs.');
+			expect(logger.error).toHaveBeenCalled();
+		});
+
+		test('propagates Storyblok eligibility failures instead of returning an empty list', async () => {
+			const { service, findMany, getPrograms } = createService();
+			const failure: ServiceResult<never> = {
+				success: false,
+				error: 'Failed to fetch programs: {"message":"down"}',
+			};
+			getPrograms.mockResolvedValue(failure);
+
+			const result = await service.getEligibleProgramsForPublicSubmission('en');
+
+			expect(result).toEqual(failure);
+			expect(findMany).not.toHaveBeenCalled();
+		});
+	});
+
 	describe('getEligibleProgramOptions', () => {
 		test('maps country, recipient count, and focuses for matching published slugs', async () => {
 			const { service, findMany } = createService();
